+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Print version information with `APP -v` and `APP version`
 -   Support on Windows operating systems
 -   Add PHP extension `calendar` as requirement
+-   Log: Do not colorize "info" events
+-   Log: Print the following events as output to STDOUT with `printAsOutput()`
+-   Log: Print the following events as messages to STDERR with `printAsMessage()`
+-   Log: Print empty line with `printEmptyLine()`
+-   Log: Allow method chaining
+-   Log: Control text formatting with tags: `<strong>`, `<u>`, `<dim>`, `<fatal>`, `<error>`, `<warning>`, `<notice>`, `<debug>`, `<red>`, `<yellow>`, `<green>`, `<grey>`
 
 
 ### Fixed